# src/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "This is a simple NodeJS API code golf project to demonstrate NodeJS, Typescript, and OpenAPI.",
		"version": "1.0.0",
		"title": "NodeChat"
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"UserModel": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobilephone": {
						"type": "string"
					},
					"loginid": {
						"type": "string"
					}				
				}
			},
			"MessageModel": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"nullable": true
					},
					"userid": {
						"type": "string",
						"nullable": true
					},
					"messagedate": {
						"type": "string",
						"nullable": true
					},
					"messagetext": {
						"type": "string",
						"nullable": true
					}				
				}
			},
			"LoginModel": {
				"type": "object",
				"properties": {
					"login": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				}		
			},
			"GeneralInvalidResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				}
			}			
		}
	},
	"servers": [{
		"url": "http://localhost:3000/",
		"description": "Development Server"
	}],
	"paths": {
		"/users": {
			"parameters": [],
			"get": {
				"description": "Get all Users",
				"security": {
					"bearerAuth": []
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/UserModel"
									}
								}								
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"parameters": [{
				"name": "id",
				"in": "path",
				"description": "id of user to return",
				"required": true,
				"type": "string"
			}],			
			"get": {
				"description": "Get Users By Id",
				"security": {
					"bearerAuth": []
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"$ref": "#/components/schemas/UserModel"
								}								
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					}
				}
			}
		},
		"/messages": {
			"parameters": [],
			"get": {
				"description": "Get all Messages",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/MessageModel"
									}
								}								
							}
						}
					}
				}
			},
			"post": {
				"description": "Create new Message",
				"parameters": [{
					"name": "body",
					"in": "body",
					"description": "Message model",
					"required": true,
					"schema": {
						"type": "object",
						"$ref": "#/components/schemas/MessageModel"
					}
				}],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"$ref": "#/components/schemas/MessageModel"
								}								
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}

					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					}										
				}
			}
		},
		"/login": {
			"post": {
				"description": "Authenticate With Service",
				"parameters": [{
					"name": "body",
					"in": "body",
					"description": "Login model",
					"required": true,
					"schema": {
						"type": "object",
						"$ref": "#/components/schemas/LoginModel"
					}
				}],
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/html": {
								"schema": {
									"type": "string"
								}								
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/html": {
								"schema": {
									"$ref": "#/components/schemas/GeneralInvalidResponse"
								}								
							}
						}
					}
				}
			}
		}
	}
}